tests/wpt/web-platform-tests/trusted-types/outerHTML.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./support/helper.js"></script>
<body>
<div id="container"></div>
<script>
  var container = document.querySelector('#container')
  test(t => {
    var html = TrustedHTML.escape(STRINGS.unescapedHTML);

    var d = document.createElement('div');
    document.querySelector('#container').appendChild(d);
    d.outerHTML = html;
    assert_equals(container.innerText, STRINGS.unescapedHTML);
    while (container.firstChild)
      container.firstChild.remove();
  }, "outerHTML = TrustedHTML.escape().");

  test(t => {
    var html = TrustedHTML.unsafelyCreate(STRINGS.unescapedHTML);

    var d = document.createElement('div');
    container.appendChild(d);
    d.outerHTML = html;
    assert_equals(container.innerText, STRINGS.unescapedText);
    while (container.firstChild)
      container.firstChild.remove();
  }, "outerHTML = TrustedHTML.unsafelyCreate().");
</script>
